Refactoring PR (#10)
* refactoring PR

* fixed typo in README

* removed newlines at beginning
