+++ conda.yaml
@@ -6,8 +6,7 @@ dependencies:
   - pandas=0.22.0
   - scikit-learn==0.19.1
   - pip:
+    - pyarrow
     - mlflow
-    - pyarrow==0.9.0.post1
-    - pyspark==2.3.0
     - xgboost==0.71
     - tensorflow==1.8.0
